+++ package.json
@@ -2,7 +2,7 @@
   "name": "BigAssFansAPI",
   "version": "1.1.1",
   "description": "An unofficial API for BigAssFans",
-  "main": "BigAssApi.js",
+  "main": "src/BigAssApi.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/sean9keenan/BigAssFansAPI.git"
